fix: change layout height, move hero buttons and text to the separate … (#32)
* fix: change layot height, move hero buttons and text to the separate layer, add padding to the content og the page

since footer is not fixed to the bottom change the wrapper min-hights property. moved the hero buttons and text to the separate layer according to the mentor remark. Add paddings to the contenct block to make the same padding in the entire app - 16px

* fix: delete content paddings

delete paddings since they affect entire page, not the just the main block

* fix: address mentor's comments

* style: add container to the carousel content, change container max-width